[0.3.0.35] Change index_t to int to fix bug

+++ gtb-types.h
@@ -11,7 +11,7 @@ typedef unsigned char 		pc_t;
 typedef uint32_t			mv_t;
 typedef long int			tbkey_t;
 typedef unsigned short  	dtm_t;
-typedef size_t 				index_t;
+typedef int 				index_t;
 
 /*<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<*/
 #ifdef __cplusplus
